requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
selenium==4.17.2
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4
python-dotenv==1.0.1
pydantic==2.6.1
sqlalchemy==2.0.25
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.3
aiosmtplib==2.0.2
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
emails==0.6
chart.js==3.9.1 